move routes-rex/analysis.ts to server/ to align with web-server zone
analysis.ts was pinned to the web-server zone by sourcevision but lived
in routes-rex/. Moved to packages/web/src/server/routes-rex-analysis.ts,
updated its relative imports, and updated the import in routes-rex/index.ts.
