+++ packages/web/src/server/routes-rex-analysis.ts
@@ -11,15 +11,15 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { randomUUID } from "node:crypto";
 import { exec as foundationExec } from "@n-dx/llm-client";
-import type { ServerContext } from "../types.js";
-import { jsonResponse, errorResponse, readBody } from "../types.js";
-import type { WebSocketBroadcaster } from "../websocket.js";
-import { insertChild, loadPRD, savePRD, appendLog } from "./shared.js";
+import type { ServerContext } from "./types.js";
+import { jsonResponse, errorResponse, readBody } from "./types.js";
+import type { WebSocketBroadcaster } from "./websocket.js";
+import { insertChild, loadPRD, savePRD, appendLog } from "./routes-rex/shared.js";
 
 import {
   type PRDItem,
   isPriority,
-} from "../rex-gateway.js";
+} from "./rex-gateway.js";
 
 // ---------------------------------------------------------------------------
 // Edited proposal types
